Improve dumb jump integration as a proper fallback
Before there was a macro but it failed at being lazy and it was required
whenever I had to configure a new mode
Now it works with a base mode mugu-lang that defines the minimal set of
feature (definition and xref) based on dumbjump.
If a particular language has a better implementation (or just use lsp) it can
just remap the command.